+++ net.sourceforge.pmd.eclipse.plugin/src/main/java/net/sourceforge/pmd/eclipse/ui/actions/internal/InternalRuleSetUtil.java
@@ -7,6 +7,7 @@ package net.sourceforge.pmd.eclipse.ui.actions.internal;
 
 import java.util.Collection;
 import java.util.HashSet;
+import java.util.Set;
 import java.util.regex.Pattern;
 
 import net.sourceforge.pmd.RuleSet;
@@ -27,6 +28,42 @@ public class InternalRuleSetUtil {
         return factory.createNewRuleSet(ruleSet.getName(), ruleSet.getDescription(), ruleSet.getFileName(),
                 InternalRuleSetUtil.convert(ruleSet.getFileExclusions()),
                 InternalRuleSetUtil.convert(includePatterns), ruleSet.getRules());
+    }
+
+    public static RuleSet addFileExclusions(RuleSet rs, Collection<Pattern> excludePatterns) {
+        return addExcludePatterns(rs, excludePatterns, new HashSet<Pattern>());
+    }
+
+    public static RuleSet addFileInclusions(RuleSet rs, Collection<Pattern> includePatterns) {
+        return addIncludePatterns(rs, includePatterns, new HashSet<Pattern>());
+    }
+
+    public static RuleSet addExcludePatterns(RuleSet ruleSet, Collection<Pattern> activeExclusionPatterns,
+            Collection<Pattern> buildPathExcludePatterns) {
+        Set<Pattern> newExcludePatterns = new HashSet<Pattern>(ruleSet.getFileExclusions());
+        newExcludePatterns.addAll(activeExclusionPatterns);
+        newExcludePatterns.addAll(buildPathExcludePatterns);
+        Set<Pattern> newIncludePatterns = new HashSet<Pattern>(ruleSet.getFileInclusions());
+
+        RuleSetFactory factory = RulesetsFactoryUtils.defaultFactory();
+        return factory.createNewRuleSet(ruleSet.getName(), ruleSet.getDescription(), ruleSet.getFileName(),
+                convert(newExcludePatterns),
+                convert(newIncludePatterns),
+                ruleSet.getRules());
+    }
+
+    public static RuleSet addIncludePatterns(RuleSet ruleSet, Collection<Pattern> activeInclusionPatterns,
+            Collection<Pattern> buildPathIncludePatterns) {
+        Set<Pattern> newExcludePatterns = new HashSet<Pattern>(ruleSet.getFileExclusions());
+        Set<Pattern> newIncludePatterns = new HashSet<Pattern>(ruleSet.getFileInclusions());
+        newIncludePatterns.addAll(activeInclusionPatterns);
+        newIncludePatterns.addAll(buildPathIncludePatterns);
+
+        RuleSetFactory factory = RulesetsFactoryUtils.defaultFactory();
+        return factory.createNewRuleSet(ruleSet.getName(), ruleSet.getDescription(), ruleSet.getFileName(),
+                convert(newExcludePatterns),
+                convert(newIncludePatterns),
+                ruleSet.getRules());
     }
 
     public static Collection<String> convert(Collection<Pattern> patterns) {